Process fork and exec monitoring core functionality
Monitor a Process Connector netlink socket for fork and/or exec notices,
and generate reports to STDOUT.  Optionally, also report these events for
threads in addition to thread group leaders.

For each report, attempt to obtain the process' command line from /proc,
sanitize it, and append it to the one-line report.